+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for Python 3.11.
+- Add option `store` to optionally disable storing results locally.
 
 ### Changed